function for sending survey question set
