+++ .gitignore
@@ -2,4 +2,4 @@ node_modules/
 public/apidoc
 .DS_Store
 */.DS_Store
-
+logfile
