+++ app.js
@@ -9,27 +9,8 @@ var index = require('./routes/index')
 var users = require('./routes/users')
 var config = require('config')
 var serverSetting = config.get('server')
-var {Pool} = require('pg')
-var dbSetting = config.get('db')
-rename(dbSetting)
 
 var app = express()
-var pool = new Pool(dbSetting)
-pool.on('error', (err, client) => {
-  console.error('Unexpected error on idle client', err)
-  process.exit(-1)
-})
-pool.connect((err, client, done) => {
-  if (err) throw err
-  client.query('SELECT * FROM hpbdata', (err, res) => {
-    done()
-    if (err) {
-      console.log(err.stack)
-    } else {
-      console.log(res.rows[0])
-    }
-  })
-})
 
 app.set('host', (process.env.HOST || serverSetting.host))
 app.set('port', (process.env.PORT || serverSetting.port))
@@ -51,18 +32,6 @@ app.use('/users', users)
 
 app.listen(app.get('port'), function () {
   console.log('Node app is running at localhost:' + app.get('port'))
-  pool.connect((err, client, release) => {
-    if (err) {
-      return console.error('Error acquiring client', err.stack)
-    }
-    client.query('SELECT NOW()', (err, result) => {
-      release()
-      if (err) {
-        return console.error('Error executing query', err.stack)
-      }
-      console.log(result.rows)
-    })
-  })
 })
 
 // catch 404 and forward to error handler
@@ -84,8 +53,3 @@ app.use(function (err, req, res, next) {
 })
 
 module.exports = app
-
-function rename (dbSetting) {
-  dbSetting['password'] = dbSetting['pAsswOrd']
-  delete dbSetting['pAsswOrd']
-}
