+++ public/psql/survey1.sql
@@ -1,0 +1,8 @@
+INSERT INTO survey_questions (qn_set, display_num, description, ans_type)
+VALUES
+ (1, 0, 'Please check the box to describe how much the following questions apply to you.', 'NONE'),
+ (1, 1, 'I am familiar with the food listed above.', '5_scale'),
+ (1, 2, 'The taste score of the food is reasonable', '5_scale'),
+ (1, 3, 'The taste score of the food reflects my preference.', '5_scale'),
+ (1, 4, 'The health score of the food is understandable.', '5_scale'),
+ (1, 5, 'The health score of the food reflect my acknowledgement.', '5_scale');
